Reorganise backend `follow` test cases
